+++ Runtime/GsplatRendererImpl.cs
@@ -11,6 +11,16 @@ namespace Gsplat
 {
     public class GsplatRendererImpl
     {
+        sealed class CandidateOrderResource : ISorterResource
+        {
+            public GraphicsBuffer OrderBuffer { get; }
+            public GraphicsBuffer InputKeys => null;
+            public bool Initialized { get; set; }
+
+            public CandidateOrderResource(GraphicsBuffer orderBuffer) => OrderBuffer = orderBuffer;
+            public void Dispose() { }
+        }
+
         public uint SplatCount { get; private set; }
 
         MaterialPropertyBlock m_propertyBlock;
@@ -24,7 +34,15 @@ namespace Gsplat
         public GraphicsBuffer CutoutsBuffer { get; private set; }
         public GraphicsBuffer OrderSizeBuffer { get; private set; }
         public GraphicsBuffer BoundsBuffer { get; private set; }
+        public GraphicsBuffer VisibleCountBuffer { get; private set; }
+        public GraphicsBuffer SortDispatchArgs { get; private set; }
+        public GraphicsBuffer DrawArgs { get; private set; }
         public ISorterResource SorterResource { get; private set; }
+        public bool FrustumCullingActive { get; private set; }
+
+        GraphicsBuffer m_candidateOrderBuffer;
+        CandidateOrderResource m_candidateOrderResource;
+        bool m_orderTargetsCandidateBuffer;
 
         static readonly int k_orderBuffer = Shader.PropertyToID("_OrderBuffer");
         static readonly int k_matrixM = Shader.PropertyToID("_MATRIX_M");
@@ -34,6 +52,33 @@ namespace Gsplat
         static readonly int k_shDegree = Shader.PropertyToID("_SHDegree");
         static readonly int k_brightness = Shader.PropertyToID("_Brightness");
         static readonly int k_scaleFactor = Shader.PropertyToID("_ScaleFactor");
+        static readonly int k_useVisibleCount = Shader.PropertyToID("_UseVisibleCount");
+        static readonly int k_visibleCountBuffer = Shader.PropertyToID("_VisibleCountBuffer");
+
+        static readonly int k_candidateCount = Shader.PropertyToID("_CandidateCount");
+        static readonly int k_useCandidateOrder = Shader.PropertyToID("_UseCandidateOrder");
+        static readonly int k_candidateOrderBuffer = Shader.PropertyToID("_CandidateOrderBuffer");
+        static readonly int k_depthBuffer = Shader.PropertyToID("_DepthBuffer");
+        static readonly int k_sortDispatchArgs = Shader.PropertyToID("_SortDispatchArgs");
+        static readonly int k_drawArgs = Shader.PropertyToID("_DrawArgs");
+        static readonly int k_eyeCount = Shader.PropertyToID("_EyeCount");
+        static readonly int k_viewportSize = Shader.PropertyToID("_ViewportSize");
+        static readonly int k_matrixMv0 = Shader.PropertyToID("_MatrixMV0");
+        static readonly int k_matrixMv1 = Shader.PropertyToID("_MatrixMV1");
+        static readonly int k_matrixMvDepth = Shader.PropertyToID("_MatrixMVDepth");
+        static readonly int k_matrixP0 = Shader.PropertyToID("_MatrixP0");
+        static readonly int k_matrixP1 = Shader.PropertyToID("_MatrixP1");
+        static readonly int k_indexCountPerInstance = Shader.PropertyToID("_IndexCountPerInstance");
+        static readonly int k_startIndex = Shader.PropertyToID("_StartIndex");
+        static readonly int k_baseVertex = Shader.PropertyToID("_BaseVertex");
+        static readonly int k_positionBuffer = Shader.PropertyToID("_PositionBuffer");
+        static readonly int k_scaleBuffer = Shader.PropertyToID("_ScaleBuffer");
+        static readonly int k_rotationBuffer = Shader.PropertyToID("_RotationBuffer");
+        static readonly int k_packedSplatsBuffer = Shader.PropertyToID("_PackedSplatsBuffer");
+
+        int m_kernelCullClear = -1;
+        int m_kernelCull = -1;
+        int m_kernelBuildArgs = -1;
 
         uint m_framesBeforeRecomputeSort = 0;
         uint m_sortsBeforeRecomputeCutouts = 0;
@@ -90,6 +135,21 @@ namespace Gsplat
             return count[0];
         }
 
+        public void SetFrustumCulling(bool active)
+        {
+            if (FrustumCullingActive == active)
+                return;
+
+            if (active)
+                EnsureCullingResources();
+            FrustumCullingActive = active;
+            SorterResource.Initialized = false;
+            if (m_candidateOrderResource != null)
+                m_candidateOrderResource.Initialized = false;
+            m_prevSplatCount = uint.MaxValue;
+            ForceRefresh();
+        }
+
         public void DispatchInitOrder(GsplatCutout[] cutouts, Matrix4x4 matrixWorld, bool cutoutsUpdateBounds)
         {
             if (cutouts.Length == 0)
@@ -118,7 +178,8 @@ namespace Gsplat
                         cutoutsUnchanged = false;
             }
 
-            if (cutoutsUnchanged && m_prevSplatCount == GsplatResource.UploadedCount)
+            if (cutoutsUnchanged && m_prevSplatCount == GsplatResource.UploadedCount &&
+                m_orderTargetsCandidateBuffer == FrustumCullingActive)
                 return;
 
             m_prevSplatCount = GsplatResource.UploadedCount;
@@ -126,9 +187,72 @@ namespace Gsplat
             CutoutsBuffer = m_gsplatAsset.UpdateCutoutsBuffer(CutoutsBuffer, m_cutoutsData);
             if (cutoutsUpdateBounds)
                 m_gsplatAsset.UpdateBoundsBuffer(BoundsBuffer);
-            m_gsplatAsset.InitOrder(SorterResource, GsplatResource, cutoutsUpdateBounds);
-            m_remainingCount = ExtractOrderSize(SorterResource.OrderBuffer);
+            var orderTarget = FrustumCullingActive ? m_candidateOrderResource : SorterResource;
+            m_gsplatAsset.InitOrder(orderTarget, GsplatResource, cutoutsUpdateBounds);
+            m_remainingCount = ExtractOrderSize(orderTarget.OrderBuffer);
             m_bounds = cutoutsUpdateBounds ? ExtractBounds() : m_gsplatAsset.Bounds;
+            m_orderTargetsCandidateBuffer = FrustumCullingActive;
+        }
+
+        public void Cull(CommandBuffer cmd, Camera camera, Transform transform)
+        {
+            var cs = m_gsplatAsset.GsplatMaterial.FrustumCullShader;
+
+            bool stereo = camera.stereoEnabled;
+            Matrix4x4 matrixM = transform.localToWorldMatrix;
+            Matrix4x4 matrixMvDepth = camera.worldToCameraMatrix * matrixM;
+            Matrix4x4 matrixMv0 = (stereo
+                ? camera.GetStereoViewMatrix(Camera.StereoscopicEye.Left)
+                : camera.worldToCameraMatrix) * matrixM;
+            Matrix4x4 matrixP0 = GL.GetGPUProjectionMatrix(stereo
+                ? camera.GetStereoProjectionMatrix(Camera.StereoscopicEye.Left)
+                : camera.projectionMatrix, false);
+            Matrix4x4 matrixMv1 = stereo
+                ? camera.GetStereoViewMatrix(Camera.StereoscopicEye.Right) * matrixM
+                : matrixMv0;
+            Matrix4x4 matrixP1 = stereo
+                ? GL.GetGPUProjectionMatrix(camera.GetStereoProjectionMatrix(Camera.StereoscopicEye.Right), false)
+                : matrixP0;
+
+            cmd.SetComputeIntParam(cs, k_candidateCount, (int)m_remainingCount);
+            cmd.SetComputeIntParam(cs, k_useCandidateOrder, m_cutoutsData.Length > 0 ? 1 : 0);
+            cmd.SetComputeIntParam(cs, k_eyeCount, stereo ? 2 : 1);
+            cmd.SetComputeVectorParam(cs, k_viewportSize,
+                new Vector4(Math.Max(1, camera.pixelWidth), Math.Max(1, camera.pixelHeight), 0, 0));
+            cmd.SetComputeMatrixParam(cs, k_matrixMv0, matrixMv0);
+            cmd.SetComputeMatrixParam(cs, k_matrixMv1, matrixMv1);
+            cmd.SetComputeMatrixParam(cs, k_matrixMvDepth, matrixMvDepth);
+            cmd.SetComputeMatrixParam(cs, k_matrixP0, matrixP0);
+            cmd.SetComputeMatrixParam(cs, k_matrixP1, matrixP1);
+            cmd.SetComputeBufferParam(cs, m_kernelCullClear, k_visibleCountBuffer, VisibleCountBuffer);
+            cmd.DispatchCompute(cs, m_kernelCullClear, 1, 1, 1);
+
+            if (m_remainingCount > 0)
+            {
+                cmd.SetComputeBufferParam(cs, m_kernelCull, k_candidateOrderBuffer, m_candidateOrderBuffer);
+                cmd.SetComputeBufferParam(cs, m_kernelCull, k_orderBuffer, SorterResource.OrderBuffer);
+                cmd.SetComputeBufferParam(cs, m_kernelCull, k_depthBuffer, SorterResource.InputKeys);
+                cmd.SetComputeBufferParam(cs, m_kernelCull, k_visibleCountBuffer, VisibleCountBuffer);
+                if (GsplatResource is GsplatResourceUncompressed uncompressed)
+                {
+                    cmd.SetComputeBufferParam(cs, m_kernelCull, k_positionBuffer, uncompressed.PositionBuffer);
+                    cmd.SetComputeBufferParam(cs, m_kernelCull, k_scaleBuffer, uncompressed.ScaleBuffer);
+                    cmd.SetComputeBufferParam(cs, m_kernelCull, k_rotationBuffer, uncompressed.RotationBuffer);
+                }
+                else if (GsplatResource is GsplatResourceSpark spark)
+                    cmd.SetComputeBufferParam(cs, m_kernelCull, k_packedSplatsBuffer, spark.PackedSplatsBuffer);
+                cmd.DispatchCompute(cs, m_kernelCull, (int)GsplatUtils.DivRoundUp(m_remainingCount, 256), 1, 1);
+            }
+
+            var mesh = GsplatSettings.Instance.Mesh;
+            cmd.SetComputeIntParam(cs, k_splatInstanceSize, (int)GsplatSettings.Instance.SplatInstanceSize);
+            cmd.SetComputeIntParam(cs, k_indexCountPerInstance, (int)mesh.GetIndexCount(0));
+            cmd.SetComputeIntParam(cs, k_startIndex, (int)mesh.GetIndexStart(0));
+            cmd.SetComputeIntParam(cs, k_baseVertex, (int)mesh.GetBaseVertex(0));
+            cmd.SetComputeBufferParam(cs, m_kernelBuildArgs, k_visibleCountBuffer, VisibleCountBuffer);
+            cmd.SetComputeBufferParam(cs, m_kernelBuildArgs, k_sortDispatchArgs, SortDispatchArgs);
+            cmd.SetComputeBufferParam(cs, m_kernelBuildArgs, k_drawArgs, DrawArgs);
+            cmd.DispatchCompute(cs, m_kernelBuildArgs, 1, 1, 1);
         }
 
         public void BindGsplatAsset(GsplatAsset gsplatAsset, bool asyncUpload = false)
@@ -138,6 +262,8 @@ namespace Gsplat
             m_gsplatAsset = gsplatAsset;
             GsplatResource = GsplatResourceManager.Get(gsplatAsset);
             gsplatAsset.SetupMaterialPropertyBlock(m_propertyBlock, GsplatResource);
+            if (FrustumCullingActive)
+                CacheCullingKernels();
             if (asyncUpload)
                 gsplatAsset.UploadDataAsync(GsplatResource);
             else
@@ -166,15 +292,51 @@ namespace Gsplat
         {
             m_propertyBlock ??= new MaterialPropertyBlock();
             m_propertyBlock.SetBuffer(k_orderBuffer, OrderBuffer);
+            if (VisibleCountBuffer != null)
+                m_propertyBlock.SetBuffer(k_visibleCountBuffer, VisibleCountBuffer);
+        }
+
+        void EnsureCullingResources()
+        {
+            if (VisibleCountBuffer != null)
+                return;
+
+            m_candidateOrderBuffer = new GraphicsBuffer(GraphicsBuffer.Target.Append, (int)SplatCount, sizeof(uint));
+            m_candidateOrderResource = new CandidateOrderResource(m_candidateOrderBuffer);
+            VisibleCountBuffer = new GraphicsBuffer(GraphicsBuffer.Target.Raw, 1, sizeof(uint));
+            SortDispatchArgs = new GraphicsBuffer(GraphicsBuffer.Target.IndirectArguments, 1, sizeof(uint) * 3);
+            DrawArgs = new GraphicsBuffer(GraphicsBuffer.Target.IndirectArguments, 1, sizeof(uint) * 5);
+            DrawArgs.SetData(new uint[5]);
+            m_propertyBlock.SetBuffer(k_visibleCountBuffer, VisibleCountBuffer);
+
+            CacheCullingKernels();
+        }
+
+        void CacheCullingKernels()
+        {
+            var cs = m_gsplatAsset.GsplatMaterial.FrustumCullShader;
+            m_kernelCullClear = cs.FindKernel("Clear");
+            m_kernelCull = cs.FindKernel(GsplatResource is GsplatResourceSpark ? "CullSpark" : "CullUncompressed");
+            m_kernelBuildArgs = cs.FindKernel("BuildArgs");
         }
 
         public void Dispose()
         {
+            FrustumCullingActive = false;
             ReleaseGsplatAsset();
             OrderBuffer?.Dispose();
             OrderBuffer = null;
             SorterResource?.Dispose();
             SorterResource = null;
+            m_candidateOrderBuffer?.Dispose();
+            m_candidateOrderBuffer = null;
+            m_candidateOrderResource = null;
+            VisibleCountBuffer?.Dispose();
+            VisibleCountBuffer = null;
+            SortDispatchArgs?.Dispose();
+            SortDispatchArgs = null;
+            DrawArgs?.Dispose();
+            DrawArgs = null;
             CutoutsBuffer?.Dispose();
             CutoutsBuffer = null;
             OrderSizeBuffer?.Dispose();
@@ -273,6 +435,7 @@ namespace Gsplat
             m_propertyBlock.SetFloat(k_brightness, brightness);
             m_propertyBlock.SetFloat(k_scaleFactor, scaleFactor);
             m_propertyBlock.SetMatrix(k_matrixM, transform.localToWorldMatrix);
+            m_propertyBlock.SetInteger(k_useVisibleCount, FrustumCullingActive ? 1 : 0);
 
             uint order = Math.Clamp(renderOrder, 0, GsplatSettings.Instance.MaxRenderOrder - 1);
             var rp = new RenderParams(m_gsplatAsset.Materials[order])
@@ -282,8 +445,14 @@ namespace Gsplat
                 layer = layer
             };
 
-            Graphics.RenderMeshPrimitives(rp, GsplatSettings.Instance.Mesh, 0,
-                Mathf.CeilToInt(m_remainingCount / (float)GsplatSettings.Instance.SplatInstanceSize));
+            if (FrustumCullingActive)
+            {
+                rp.worldBounds = new Bounds(transform.position, Vector3.one * 1e6f);
+                Graphics.RenderMeshIndirect(rp, GsplatSettings.Instance.Mesh, DrawArgs);
+            }
+            else
+                Graphics.RenderMeshPrimitives(rp, GsplatSettings.Instance.Mesh, 0,
+                    Mathf.CeilToInt(m_remainingCount / (float)GsplatSettings.Instance.SplatInstanceSize));
         }
     }
 }